Trait bp_wococo::WococoFinalityApi

pub trait WococoFinalityApi<Block: BlockT>: Core<Block> {
    fn best_finalized(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(BlockNumber, Hash), ApiError> { ... }

    fn best_finalized_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(BlockNumber, Hash), ApiError> { ... }

    fn is_known_header(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        hash: Hash
    ) -> Result<bool, ApiError> { ... }

    fn is_known_header_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        hash: Hash
    ) -> Result<bool, ApiError> { ... }
}

API for querying information about the finalized Wococo headers.

This API is implemented by runtimes that are bridging with the Wococo chain, not the Wococo runtime itself.

Provided methods

fn best_finalized(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(BlockNumber, Hash), ApiError>

Returns number and hash of the best finalized header known to the bridge module.

fn best_finalized_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(BlockNumber, Hash), ApiError>

Returns number and hash of the best finalized header known to the bridge module.

fn is_known_header(
&self,
__runtime_api_at_param__: &BlockId<Block>,
hash: Hash
) -> Result<bool, ApiError>

Returns true if the header is known to the runtime.

fn is_known_header_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
hash: Hash
) -> Result<bool, ApiError>

Returns true if the header is known to the runtime.

Loading content...